+++ src/ble/att_db.h
@@ -48,12 +48,15 @@ extern "C" {
 #endif
 
 // custom BTstack error codes
-#define ATT_ERROR_HCI_DISCONNECT_RECEIVED          0x1f
+#define ATT_ERROR_HCI_DISCONNECT_RECEIVED         0x1f
 
 // custom BTstack ATT error codes
-#define ATT_ERROR_DATA_MISMATCH                    0x7e
-#define ATT_ERROR_TIMEOUT                          0x7F
-    
+#define ATT_ERROR_DATA_MISMATCH                   0x7e
+#define ATT_ERROR_TIMEOUT                         0x7F
+
+// custom BTstack ATT Response Pending 
+#define ATT_READ_RESPONSE_PENDING                 0xffff
+
 typedef struct att_connection {
     hci_con_handle_t con_handle;
     uint16_t mtu;       // initialized to ATT_DEFAULT_MTU (23), negotiated during MTU exchange
@@ -66,6 +69,7 @@ typedef struct att_connection {
 // ATT Client Read Callback for Dynamic Data
 // - if buffer == NULL, don't copy data, just return size of value
 // - if buffer != NULL, copy data and return number bytes copied
+// If ENABLE_ATT_DELAYED_READ_RESPONSE is defined, you may return ATT_READ_RESPONSE_PENDING if data isn't available yet
 // @param con_handle of hci le connection
 // @param attribute_handle to be read
 // @param offset defines start of attribute value
@@ -130,7 +134,8 @@ void att_dump_attributes(void);
  * @param att_connection used for mtu and security properties
  * @param request_buffer, request_len: ATT request from clinet
  * @param response_buffer for result
- * @returns len of data in response buffer. 0 = no response
+ * @returns len of data in response buffer. 0 = no response, 
+ *          ATT_READ_RESPONSE_PENDING if it was returned at least once for dynamic data (requires ENABLE_ATT_DELAYED_READ_RESPONSE)
  */
 uint16_t att_handle_request(att_connection_t * att_connection,
                             uint8_t * request_buffer,