+++ src/ble/le_device_db_tlv.c
@@ -170,7 +170,14 @@ int le_device_db_count(void){
 	return num_valid_entries;
 }
 
+int le_device_db_max_count(void){
+    return NVM_NUM_DEVICE_DB_ENTRIES;
+}
+
 void le_device_db_remove(int index){
+    // check if entry exists
+    if (entry_map[index] == 0) return; 
+
 	// delete entry in TLV
 	le_device_db_tlv_delete(index);