+++ CHANGELOG.md
@@ -7,17 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+ATT Server: support for delayed ATT read response, see example/att_delayed_read_response.c
 
 ### Changed
 - GAP: allow to limit number of connections in LE Peripheral role with gap_set_max_number_peripheral_connections
 
 ### Fixed
+- RFCOMM: fix infinite loop on L2CAP connection error (regression from 4c3eeed1)
+- HSP HS: accept incomming SCO connection
+- SM: fix iteration over LE Device DB entries for bonding and address resolving
 
 ## Changes February 2018
 
 ### Added
-- Port for Windows with Zephyr HCI Firmware connected via serial port  
-- em9304: ability to upload patch containers during HCI bootup.
 - GATT Client: gatt_client_request_can_write_without_response_event() causes GATT_EVENT_CAN_WRITE_WITHOUT_RESPONSE
 - SM: new event SM_EVENT_PAIRING_COMPLETE
 - GAP: support iteration over stored Classic link keys: gap_link_key_iterator_init, gap_link_key_iterator_get_next, gap_link_key_iterator_done